agentes-ia/whatsapp-webhook/requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-cloud-secret-manager
google-cloud-logging
